+++ app/src/main/kotlin/com/omnitune/app/constants/PreferenceKeys.kt
@@ -31,6 +31,8 @@ val UseNewLibraryDesignKey = booleanPreferencesKey("useNewLibraryDesign")
 val UseNewMiniPlayerDesignKey = booleanPreferencesKey("useNewMiniPlayerDesign")
 val HidePlayerThumbnailKey = booleanPreferencesKey("hidePlayerThumbnail")
 val OmniTuneCanvasKey = booleanPreferencesKey("veluneCanvas")
+val PlaybackQualityModeKey = stringPreferencesKey("playbackQualityMode")
+
 val ThumbnailCornerRadiusKey = floatPreferencesKey("thumbnailCornerRadius")
 val CropThumbnailToSquareKey = booleanPreferencesKey("cropThumbnailToSquare")
 val SeekExtraSeconds = booleanPreferencesKey("seekExtraSeconds")